Moved storage type to avatar package.
