+++ src/main/java/one/stayfocused/backend/service/avatar/AvatarStorageFactory.java
@@ -1,6 +1,5 @@
 package one.stayfocused.backend.service.avatar;
 
-import one.stayfocused.backend.storage.StorageType;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
