+++ src/main/java/one/stayfocused/backend/service/avatar/AvatarStorageService.java
@@ -1,6 +1,5 @@
 package one.stayfocused.backend.service.avatar;
 
-import one.stayfocused.backend.storage.StorageType;
 import org.springframework.web.multipart.MultipartFile;
 
 public interface AvatarStorageService {
