+++ src/main/java/one/stayfocused/backend/service/avatar/StorageType.java
@@ -1,4 +1,4 @@
-package one.stayfocused.backend.storage;
+package one.stayfocused.backend.service.avatar;
 
 public enum StorageType {
     LOCAL, CLOUDINARY
